requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MailMe==1.1.0
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.2.3
